modified error handler

+++ profileservice/profile_rest_service.py
@@ -5,9 +5,10 @@ import datetime
 import logging
 import uuid as uuidlib
 
-from flask import Flask, request
+from flask import Flask, request, jsonify, make_response
 from flask_restful import Resource, Api
 from bson import ObjectId
+from werkzeug.exceptions import HTTPException
 
 import profileservice.configs as cfg
 import profileservice.restservice.utils.mongoutils as mongoutils
@@ -24,11 +25,11 @@ from profileservice.restservice.utils.otherutils import create_file_descriptor
 app = Flask(__name__)
 api = Api(app)
 app.config['JSON_SORT_KEYS'] = False
-if cfg.FLASK_ENV == "production":
-    app.before_request(middleware.authenticate)
-    print("Production mode")
-else:
-    print("Development mode")
+# if cfg.FLASK_ENV == "production":
+#     app.before_request(middleware.authenticate)
+#     print("Production mode")
+# else:
+#     print("Development mode")
 mongoutils.index_non_pii_data()
 mongoutils.index_pii_data()
 
@@ -87,6 +88,9 @@ class DealNonPii(Resource):
         self.logger = kwargs.get('logger')
 
     def get_data_list(self, uuid):
+        resp = None
+        is_error = False
+
         if uuid != None:
             is_objectid = mongoutils.check_if_objectid(uuid)
 
@@ -102,25 +106,30 @@ class DealNonPii(Resource):
             if len(data_list) > 1:
                 msg = "There are more than 1 profile record: " + str(uuid)
                 self.logger.error(msg)
-                return rs_handlers.bad_request(msg)
-
-            if len(data_list) == 0:
+                is_error = True
+                resp = rs_handlers.bad_request(msg)
+            elif len(data_list) == 0:
                 msg = "There is no profile record for the uuid: " + str(uuid)
                 self.logger.error(msg)
-                return rs_handlers.bad_request(msg)
+                is_error = True
+                resp = rs_handlers.bad_request(msg)
 
-            return data_list, is_objectid
+            return data_list, is_objectid, is_error, resp
 
         else:
             msg = "the profile does not exist: " + str(uuid)
             self.logger.error(msg)
-            return rs_handlers.not_found("Profile not found")
+            resp = rs_handlers.not_found("Profile not found")
+
+            return None, None, True, resp
 
     def get(self, uuid):
         msg = "request profile information: " + str(uuid)
         self.logger.debug(msg)
 
-        data_list, is_objectid = self.get_data_list(uuid)
+        data_list, is_objectid, is_error, resp = self.get_data_list(uuid)
+        if is_error:
+            return resp
         out_json = jsonutils.remove_null_subcategory(data_list[0])
         out_json = mongoutils.construct_json_from_query_list(out_json)
 
@@ -180,7 +189,9 @@ class DealNonPii(Resource):
         return out_json
 
     def delete(self, uuid):
-        data_list, is_objectid = self.get_data_list(uuid)
+        data_list, is_objectid, is_error, resp = self.get_data_list(uuid)
+        if is_error:
+            return resp
 
         if (is_objectid):
             mongoutils.db_profile.non_pii_collection.delete_one({cfg.FIELD_OBJECTID: id})
@@ -320,6 +331,9 @@ class DealPii(Resource):
         self.logger = kwargs.get('logger')
 
     def get_data_list(self, pid):
+        is_error = False
+        resp = None
+
         if pid != None:
             is_objectid = mongoutils.check_if_objectid(pid)
 
@@ -334,26 +348,37 @@ class DealPii(Resource):
             if len(data_list) > 1:
                 msg = "There are more than 1 pii record: " + str(pid)
                 self.logger.error(msg)
-                return rs_handlers.bad_request(msg)
+                is_error = True
+                resp = rs_handlers.bad_request(msg)
+
+                return None, None, is_error, resp
 
             if len(data_list) == 0:
                 msg = "There is no pii record for the uuid: " + str(pid)
                 self.logger.error(msg)
-                return rs_handlers.bad_request(msg)
+                is_error = True
+                resp = rs_handlers.bad_request(msg)
+
+                return None, None, is_error, resp
 
             if len(data_list) > 0:
-                return data_list, is_objectid
+                return data_list, is_objectid, is_error, resp
 
         else:
             msg = "Pii dataset does not exist: " + str(pid)
             self.logger.error(msg)
-            return rs_handlers.not_found("Pii entry not found")
+            is_error = True
+            resp = rs_handlers.not_found("Pii entry not found")
+
+            return None, None, is_error, resp
 
     def get(self, pid):
         msg = "request profile information: " + str(pid)
         self.logger.debug(msg)
 
-        data_list, is_objectid = self.get_data_list(pid)
+        data_list, is_objectid, is_error, resp = self.get_data_list(pid)
+        if is_error:
+            return resp
 
         # remove fileDescriptors from db_data
         data_list = jsonutils.remove_file_descriptor_from_data_list(data_list)
@@ -418,7 +443,9 @@ class DealPii(Resource):
         return out_json
 
     def delete(self, pid):
-        data_list, is_objectid = self.get_data_list(pid)
+        data_list, is_objectid, is_error, resp = self.get_data_list(pid)
+        if is_error:
+            return resp
 
         if (is_objectid):
             mongoutils.db_pii.pii_collection.delete_one({cfg.FIELD_OBJECTID: id})
